refactor: (#164) 빠른시작 도메인 분리 및 응용서비스를 도메인 서비스로 이관한다
